refactor: move key races pages to a specific folder
