+++ src/app/[locale]/races/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next'
 
-import { LocationUnitedStates } from '@/components/app/pageLocationUnitedStates'
+import { LocationUnitedStates } from '@/components/app/pagesKeyRaces/pageLocationUnitedStates'
 import { queryDTSILocationUnitedStatesInformation } from '@/data/dtsi/queries/queryDTSILocationUnitedStatesInformation'
 import { PageProps } from '@/types'
 import { generateMetadataDetails } from '@/utils/server/metadataUtils'
