+++ src/app/[locale]/races/presidential/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next'
 
-import { LocationRaceSpecific } from '@/components/app/pageLocationRaceSpecific'
+import { LocationRaceSpecific } from '@/components/app/pagesKeyRaces/pageLocationRaceSpecific'
 import { queryDTSILocationUnitedStatesPresidential } from '@/data/dtsi/queries/queryDTSILocationUnitedStatesPresidentialInformation'
 import { PageProps } from '@/types'
 import { generateMetadataDetails } from '@/utils/server/metadataUtils'
