+++ src/components/app/pagesKeyRaces/common/constants.ts
@@ -1,0 +1,1 @@
+export const KEY_RACES_ADDRESS_STORAGE_KEY = '@swc/key-races-address'
