+++ src/components/app/pagesKeyRaces/pageEndorsedCandidates/index.tsx
@@ -2,7 +2,7 @@ import { DarkHeroSection } from '@/components/app/darkHeroSection'
 import { DTSIPersonHeroCard } from '@/components/app/dtsiPersonHeroCard'
 import { DTSIPersonHeroCardRow } from '@/components/app/dtsiPersonHeroCard/dtsiPersonHeroCardRow'
 import { PACFooter } from '@/components/app/pacFooter'
-import { MaybeDonateButton } from '@/components/app/pageEndorsedCandidates/dtsiPersonHeroCardDonateButton'
+import { MaybeDonateButton } from '@/components/app/pagesKeyRaces/pageEndorsedCandidates/dtsiPersonHeroCardDonateButton'
 import { InternalLink } from '@/components/ui/link'
 import { PageTitle } from '@/components/ui/pageTitleText'
 import { DTSI_EndorsedCandidatesQuery } from '@/data/dtsi/generated'
