+++ src/components/app/pagesKeyRaces/pageLocationStateSpecific/userLocationRaceInfo.tsx
@@ -6,7 +6,7 @@ import { noop } from 'lodash-es'
 
 import { DTSIPersonHeroCard } from '@/components/app/dtsiPersonHeroCard'
 import { DTSIPersonHeroCardRow } from '@/components/app/dtsiPersonHeroCard/dtsiPersonHeroCardRow'
-import { organizeStateSpecificPeople } from '@/components/app/pageLocationStateSpecific/organizeStateSpecificPeople'
+import { organizeStateSpecificPeople } from '@/components/app/pagesKeyRaces/pageLocationStateSpecific/organizeStateSpecificPeople'
 import { Button } from '@/components/ui/button'
 import { GooglePlacesSelect, GooglePlacesSelectProps } from '@/components/ui/googlePlacesSelect'
 import { InternalLink } from '@/components/ui/link'
